requirements.txt
awscli
gcloud
azure-cli
